% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwise-l2-cd.R
\name{get.fold.id}
\alias{get.fold.id}
\title{Generate \code{fold.id} of Given Size and Length}
\usage{
get.fold.id(n, nfold)
}
\arguments{
\item{n}{integer value. The length of \code{fold.id}}

\item{nfold}{integer value. The number of folds.}
}
\value{
integer vector. The value should between 1 and nfold.
}
