% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel.R
\name{get.Hv}
\alias{get.Hv}
\title{Get the Hessian-Vector Multiplication Function of Gumbel Regression Given the Response and Covariates}
\usage{
get.Hv(X, y, implementation = c("r", "cpp"))
}
\arguments{
\item{X}{matrix. The covariates.}

\item{y}{numeric vector. The response.}

\item{implementation}{character value. Select the implemenation.}
}
\value{
function with two parameters: \code{w} and \code{s}.
The first element of \code{w} is the log of scale parameter of gumbel distribution.
The other elements are regression coefficients.
The \code{s} is the vector which will be multiplied by the hessian.
}
