% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel.R
\name{get.loss}
\alias{get.loss}
\title{Get the Loss Function of Gumbel Regression Given the Response and Covariates}
\usage{
get.loss(X, y, implementation = c("r", "cpp"))
}
\arguments{
\item{X}{matrix. The covariates.}

\item{y}{numeric vector. The response.}

\item{implementation}{character value. Select the implemenation.}
}
\value{
function with single parameter which should be a numeric vector \code{w}.
The first element of \code{w} is the log of scale parameter of gumbel distribution.
The other elements are regression coefficients.
}
