% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel.R
\name{get.moment}
\alias{get.moment}
\title{Moment Estimator of Gumbel Distribution}
\usage{
get.moment(y)
}
\arguments{
\item{y}{numeric vector.}
}
\value{
A list with following elements:
\itemize{
 \item mu. The location parameter.
 \item sigma. The scale parameter.
}
}
