% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwise-l2-cd.R
\name{gumbelRegression}
\alias{gumbelRegression}
\title{Pathwise L2 Regularization and Coordinate Descent}
\usage{
gumbelRegression(X, y, fold.id, lambda.seq = 10^seq(1, -4, length.out = 100),
  tolerance = 1e-04,
  implementation = getOption("gumbelRegression.implementation", "r"))
}
\arguments{
\item{X}{matrix. The covariates.}

\item{y}{numeric vector. The response.}

\item{fold.id}{integer vector. The fold id of each instances. The number should be 1 to \code{nfold}.}

\item{lambda.seq}{decreasing numeric vector. The sequence of L2 regularization.}

\item{implementation}{character value. Select the implemenation.}
}
